Fix jump addressing after 2886e2aa1.
In 2886e2aa1 the size of size of the address for the jump switched to 32 bits,
which means the brainfuck compiler needs to adjust and emit 32 bit null values
that will then be fixed up by bf2y.

The fixups done in bf2y are now calculating the relative distance between the
jump targets during the fixup phase of the setup.

Lastly, because I have implemented a fibonacii generator as a PoC change the
OP_BF_OUTPUT instruction to print an unsigned 64bit integer instead of a
character. This means helloworld.bf and cat.bf no longer work perfectly (they
now print their output in hexlified format. ;)